Refactor worksheet implementation into focused modules
