+++ kernel/src/api/worksheet/worksheet-events.ts
@@ -1,0 +1,55 @@
+export const SHEET_EVENT_TO_INTERNAL: Record<string, string[]> = {
+  cellChanged: [
+    'cell:changed',
+    'cells:batch-changed',
+    'cell:value-changed',
+    'cell:format-changed',
+    'cell:metadata-changed',
+  ],
+  filterChanged: ['filter:changed', 'filter:applied', 'filter:cleared', 'filter:column-changed'],
+  visibilityChanged: [
+    'visibility:changed',
+    'row:hidden',
+    'row:shown',
+    'column:hidden',
+    'column:shown',
+  ],
+  structureChanged: [
+    'structure:changed',
+    'row:inserted',
+    'row:deleted',
+    'rows:inserted',
+    'rows:deleted',
+    'column:inserted',
+    'column:deleted',
+    'columns:inserted',
+    'columns:deleted',
+  ],
+  mergeChanged: ['merge:changed', 'merge:created', 'merge:removed'],
+  tableChanged: ['table:changed', 'table:created', 'table:updated', 'table:deleted'],
+  chartChanged: ['chart:changed', 'chart:created', 'chart:updated', 'chart:deleted'],
+  slicerChanged: ['slicer:changed', 'slicer:created', 'slicer:updated', 'slicer:deleted'],
+  sparklineChanged: [
+    'sparkline:changed',
+    'sparkline:created',
+    'sparkline:updated',
+    'sparkline:deleted',
+  ],
+  groupingChanged: ['grouping:changed', 'grouping:rows-changed', 'grouping:columns-changed'],
+  cfChanged: [
+    'cf:changed',
+    'cf:rules-changed',
+    'cf:rule-added',
+    'cf:rule-removed',
+    'cf:rules-cleared',
+  ],
+  viewportRefreshed: ['viewport:data-changed', 'viewport:refreshed'],
+  recalcComplete: ['recalc:complete', 'calc:complete'],
+  nameChanged: ['sheet:renamed'],
+  selectionChanged: ['selection:changed'],
+  activated: ['sheet:activated'],
+  columnSorted: ['sort:column-sorted'],
+  rowSorted: ['sort:row-sorted'],
+  formulaChanged: ['formula:changed'],
+  protectionChanged: ['protection:changed'],
+};
